Create "leshan-core-cf" project.
