+++ leshan-client-cf/src/main/java/org/eclipse/leshan/client/californium/impl/BootstrapResource.java
@@ -16,6 +16,8 @@
  *******************************************************************************/
 package org.eclipse.leshan.client.californium.impl;
 
+import static org.eclipse.leshan.core.californium.ResponseCodeUtil.fromLwM2mCode;
+
 import org.eclipse.californium.core.CoapResource;
 import org.eclipse.californium.core.server.resources.CoapExchange;
 import org.eclipse.californium.core.server.resources.Resource;
@@ -23,7 +25,6 @@ import org.eclipse.leshan.client.request.ServerIdentity;
 import org.eclipse.leshan.client.servers.BootstrapHandler;
 import org.eclipse.leshan.core.request.BootstrapFinishRequest;
 import org.eclipse.leshan.core.response.BootstrapFinishResponse;
-
 /**
  * A CoAP {@link Resource} in charge of handling the Bootstrap Finish indication from the bootstrap server.
  */
@@ -40,7 +41,7 @@ public class BootstrapResource extends CoapResource {
     public void handlePOST(CoapExchange exchange) {
         ServerIdentity identity = ResourceUtil.extractServerIdentity(exchange, bootstrapHandler);
         BootstrapFinishResponse response = bootstrapHandler.finished(identity, new BootstrapFinishRequest());
-        exchange.respond(ResourceUtil.fromLwM2mCode(response.getCode()), response.getErrorMessage());
+        exchange.respond(fromLwM2mCode(response.getCode()), response.getErrorMessage());
     }
 
 }
