+++ leshan-client-cf/src/main/java/org/eclipse/leshan/client/californium/impl/LwM2mClientResponseBuilder.java
@@ -15,10 +15,10 @@
  *******************************************************************************/
 package org.eclipse.leshan.client.californium.impl;
 
-import org.eclipse.californium.core.coap.CoAP;
+import static org.eclipse.leshan.core.californium.ResponseCodeUtil.fromCoapCode;
+
 import org.eclipse.californium.core.coap.Request;
 import org.eclipse.californium.core.coap.Response;
-import org.eclipse.leshan.ResponseCode;
 import org.eclipse.leshan.core.request.BootstrapRequest;
 import org.eclipse.leshan.core.request.DeregisterRequest;
 import org.eclipse.leshan.core.request.RegisterRequest;
@@ -30,7 +30,6 @@ import org.eclipse.leshan.core.response.DeregisterResponse;
 import org.eclipse.leshan.core.response.LwM2mResponse;
 import org.eclipse.leshan.core.response.RegisterResponse;
 import org.eclipse.leshan.core.response.UpdateResponse;
-import org.eclipse.leshan.util.Validate;
 
 public class LwM2mClientResponseBuilder<T extends LwM2mResponse> implements UplinkRequestVisitor {
 
@@ -40,40 +39,6 @@ public class LwM2mClientResponseBuilder<T extends LwM2mResponse> implements Upli
     private final Response coapResponse;
 
     private LwM2mResponse lwM2mresponse;
-
-    // TODO leshan-code-cf: this code should be factorize in a leshan-core-cf project.
-    // duplicate from org.eclipse.leshan.server.californium.impl.LwM2mResponseBuilder<T>
-    public static ResponseCode fromCoapCode(final int code) {
-        Validate.notNull(code);
-
-        if (code == CoAP.ResponseCode.CREATED.value) {
-            return ResponseCode.CREATED;
-        } else if (code == CoAP.ResponseCode.DELETED.value) {
-            return ResponseCode.DELETED;
-        } else if (code == CoAP.ResponseCode.CHANGED.value) {
-            return ResponseCode.CHANGED;
-        } else if (code == CoAP.ResponseCode.CONTENT.value) {
-            return ResponseCode.CONTENT;
-        } else if (code == CoAP.ResponseCode.BAD_REQUEST.value) {
-            return ResponseCode.BAD_REQUEST;
-        } else if (code == CoAP.ResponseCode.UNAUTHORIZED.value) {
-            return ResponseCode.UNAUTHORIZED;
-        } else if (code == CoAP.ResponseCode.NOT_FOUND.value) {
-            return ResponseCode.NOT_FOUND;
-        } else if (code == CoAP.ResponseCode.METHOD_NOT_ALLOWED.value) {
-            return ResponseCode.METHOD_NOT_ALLOWED;
-        } else if (code == CoAP.ResponseCode.FORBIDDEN.value) {
-            return ResponseCode.FORBIDDEN;
-        } else if (code == CoAP.ResponseCode.UNSUPPORTED_CONTENT_FORMAT.value) {
-            return ResponseCode.UNSUPPORTED_CONTENT_FORMAT;
-        } else if (code == CoAP.ResponseCode.NOT_ACCEPTABLE.value) {
-            return ResponseCode.NOT_ACCEPTABLE;
-        } else if (code == CoAP.ResponseCode.INTERNAL_SERVER_ERROR.value) {
-            return ResponseCode.INTERNAL_SERVER_ERROR;
-        } else {
-            throw new IllegalArgumentException("Invalid CoAP code for LWM2M response: " + code);
-        }
-    }
 
     public LwM2mClientResponseBuilder(final Request coapRequest, final Response coapResponse) {
         this.coapRequest = coapRequest;
