+++ leshan-server-cf/pom.xml
@@ -31,6 +31,10 @@ Contributors:
     <dependencies>
         <dependency>
             <groupId>org.eclipse.leshan</groupId>
+            <artifactId>leshan-core-cf</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>org.eclipse.leshan</groupId>
             <artifactId>leshan-server-core</artifactId>
         </dependency>
         <dependency>
