+++ leshan-server-cf/src/main/java/org/eclipse/leshan/server/californium/impl/LwM2mResponseBuilder.java
@@ -15,7 +15,8 @@
  *******************************************************************************/
 package org.eclipse.leshan.server.californium.impl;
 
-import org.eclipse.californium.core.coap.CoAP;
+import static org.eclipse.leshan.core.californium.ResponseCodeUtil.fromCoapCode;
+
 import org.eclipse.californium.core.coap.MediaTypeRegistry;
 import org.eclipse.californium.core.coap.Request;
 import org.eclipse.californium.core.coap.Response;
@@ -54,7 +55,6 @@ import org.eclipse.leshan.core.response.ReadResponse;
 import org.eclipse.leshan.core.response.WriteAttributesResponse;
 import org.eclipse.leshan.core.response.WriteResponse;
 import org.eclipse.leshan.server.client.Registration;
-import org.eclipse.leshan.util.Validate;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -69,40 +69,6 @@ public class LwM2mResponseBuilder<T extends LwM2mResponse> implements DownlinkRe
     private final Registration registration;
     private final LwM2mModel model;
     private final LwM2mNodeDecoder decoder;
-
-    // TODO leshan-code-cf: this code should be factorize in a leshan-core-cf project.
-    // duplicate from org.eclipse.leshan.client.californium.impl.LwM2mClientResponseBuilder<T>
-    public static ResponseCode fromCoapCode(final int code) {
-        Validate.notNull(code);
-
-        if (code == CoAP.ResponseCode.CREATED.value) {
-            return ResponseCode.CREATED;
-        } else if (code == CoAP.ResponseCode.DELETED.value) {
-            return ResponseCode.DELETED;
-        } else if (code == CoAP.ResponseCode.CHANGED.value) {
-            return ResponseCode.CHANGED;
-        } else if (code == CoAP.ResponseCode.CONTENT.value) {
-            return ResponseCode.CONTENT;
-        } else if (code == CoAP.ResponseCode.BAD_REQUEST.value) {
-            return ResponseCode.BAD_REQUEST;
-        } else if (code == CoAP.ResponseCode.UNAUTHORIZED.value) {
-            return ResponseCode.UNAUTHORIZED;
-        } else if (code == CoAP.ResponseCode.NOT_FOUND.value) {
-            return ResponseCode.NOT_FOUND;
-        } else if (code == CoAP.ResponseCode.METHOD_NOT_ALLOWED.value) {
-            return ResponseCode.METHOD_NOT_ALLOWED;
-        } else if (code == CoAP.ResponseCode.FORBIDDEN.value) {
-            return ResponseCode.FORBIDDEN;
-        } else if (code == CoAP.ResponseCode.UNSUPPORTED_CONTENT_FORMAT.value) {
-            return ResponseCode.UNSUPPORTED_CONTENT_FORMAT;
-        } else if (code == CoAP.ResponseCode.NOT_ACCEPTABLE.value) {
-            return ResponseCode.NOT_ACCEPTABLE;
-        } else if (code == CoAP.ResponseCode.INTERNAL_SERVER_ERROR.value) {
-            return ResponseCode.INTERNAL_SERVER_ERROR;
-        } else {
-            throw new IllegalArgumentException("Invalid CoAP code for LWM2M response: " + code);
-        }
-    }
 
     public LwM2mResponseBuilder(final Request coapRequest, final Response coapResponse, final Registration registration,
             final LwM2mModel model, final ObservationServiceImpl observationService,
